+++ src/main/java/com/example/setting/controller/CommentController.java
@@ -1,0 +1,2 @@
+package com.example.setting.controller;public class CommentController {
+}
