+++ src/main/java/com/example/setting/dto/CommentDTO.java
@@ -1,0 +1,19 @@
+package com.example.setting.dto;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class ReplyDto {
+    private String content;
+
+    public ReplyDto() {
+    }
+
+    public ReplyDto(String content) {
+        this.content = content;
+    }
+
+    //
+}
