+++ src/main/java/com/example/setting/service/CommentService.java
@@ -1,0 +1,2 @@
+package com.example.setting.service;public class CommentService {
+}
